Don't install methods in FALLBACK
While this has sort of worked in the past (however, would have been vulnerable to crashes in the case of parallelism), it depends on an implementation detail of method caching. Upcoming Rakudo changes in this area thus result in a test failure for this module. Furthermore, since the same changes also will allow for calls made to `FALLBACK` to be as efficient as an installed method, any performance benefits of installing a method in `FALLBACK` also no longer apply.